Gestione docker + fix db

+++ deamon/src/main/java/cybersec/deception/deamon/utils/DatabaseUtils.java
@@ -1,6 +1,7 @@
 package cybersec.deception.deamon.utils;
 
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
@@ -12,13 +13,14 @@ import java.sql.Statement;
 import java.util.List;
 import java.util.Map;
 
+@Component
 public class DatabaseUtils {
 
-    private static String driverClass;
-    private static String url;
-    private static String username;
-    private static String password;
-    private static String dbname;
+    public static String driverClass;
+    public static String url;
+    public static String username;
+    public static String password;
+    public static String dbname;
 
     @Value("${database.driverclass}")
     public void setDriverClass(String driverClass) {
@@ -64,19 +66,33 @@ public class DatabaseUtils {
             // Seleziona il database
             statement.executeUpdate("USE " + dbname);
 
-            // Modifico lo sciprt .sql in base alle entità e poi lo eseguo
-            Statement finalStatement = statement;
-            entities.forEach((key, value) -> {
-                    String outputFilePath = SQLFilesUtils.getUpdatedSqlFile(key, value);
-                    executeSqlScript(finalStatement, outputFilePath);
-                    FileUtils.deleteFile(outputFilePath);
-                }
-            );
+            if (true) {
+                String sqlScript = readScriptFromFile("deamon/src/main/resources/static/templates/UserCreate.sql");
+                statement.executeUpdate(sqlScript);
+
+                String sqlddScript = readScriptFromFile("deamon/src/main/resources/static/templates/User.sql");
+                statement.executeUpdate(sqlddScript);
+            }
+            else {
+                // Modifico lo sciprt .sql in base alle entità e poi lo eseguo
+                Statement finalStatement = statement;
+                entities.forEach((key, value) -> {
+                            String outputFilePath = SQLFilesUtils.getUpdatedSqlFile(key, value);
+                            if (!Utils.isNullOrEmpty(outputFilePath)) {
+                                executeSqlScript(finalStatement, outputFilePath);
+                                FileUtils.deleteFile(outputFilePath);
+                            }
+                        }
+                );
+            }
+
 
             System.out.println("Dati popolati con successo");
 
         } catch (ClassNotFoundException | SQLException e) {
             e.printStackTrace();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
         } finally {
             try {
                 if (statement != null) {
@@ -91,7 +107,24 @@ public class DatabaseUtils {
         }
     }
 
-    public static void executeSqlScript(Statement statement, String filePath) {
+    private static String readScriptFromFile(String filePath) throws IOException {
+        StringBuilder scriptContent = new StringBuilder();
+        try (BufferedReader reader = new BufferedReader(new FileReader(filePath))) {
+            String line;
+            while ((line = reader.readLine()) != null) {
+                // Aggiungi la riga al contenuto dello script
+                scriptContent.append(line).append("\n");
+            }
+        }
+        return scriptContent.toString();
+    }
+
+    private static List<String> readInsertToScript() throws IOException {
+
+        return null;
+    }
+
+    private static void executeSqlScript(Statement statement, String filePath) {
         try {
 
             // Leggi il file SQL
